+++ src/main/java/ladder/domain/LadderLine.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public class LadderLine {
