+++ src/main/java/ladder/domain/LadderLinePoint.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public class LadderLinePoint {
