+++ src/main/java/ladder/domain/LadderLinePoints.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 import java.util.Collections;
 import java.util.List;
