+++ src/test/java/ladder/LadderHeightTest.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.domain.LadderHeight;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
