+++ src/test/java/ladder/LadderLinePainterTest.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.domain.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
