+++ src/test/java/ladder/LadderLinePointPainterTest.java
@@ -1,5 +1,8 @@
 package ladder;
 
+import ladder.domain.LadderLinePoint;
+import ladder.domain.LadderLinePointPainter;
+import ladder.domain.LadderLinePoints;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
