+++ src/test/java/ladder/LadderLinePointsTest.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import ladder.domain.LadderLinePoint;
+import ladder.domain.LadderLinePoints;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
